+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Add missing return types annotations
 * Improve the WITH statements parser (#363)
 * Add support for passing `Context::SQL_MODE*` constants to `Context::setMode` method
+* Fix additional body tokens issue with `CREATE VIEW` statements (#371)
 
 ## [5.5.0] - 2021-12-08